test: restructure spec tree to mirror lua/ layout
Four categories of changes; net -356 lines, 85 -> 83 spec files, and
the framework can now parallelize the previously monolithic
explorer_spec.lua across two workers (wall-clock 35.0s -> 28.4s in
a full local run).

A. Move specs whose source is elsewhere:
   tests/ui/scrollsync_spec.lua        -> tests/scrollsync_spec.lua
      (source: lua/codediff/scrollsync.lua, top-level not under ui/)
   tests/ui/keymap/                    -> tests/keymap/
      (source: lua/codediff/keymap/, top-level not under ui/)
   tests/flatten_dirs_spec.lua         -> tests/ui/explorer/flatten_dirs_spec.lua
      (tests an explorer feature; belongs with the other explorer specs)
   tests/ui/core_spec.lua              -> tests/ui/render_core_spec.lua
      (source: lua/codediff/ui/core.lua; renamed so it no longer collides
      mentally with tests/core/ which covers lua/codediff/core/*)

B. Split explorer_spec.lua at its natural describe seam:
   explorer_spec.lua                   -> explorer_spec.lua (17 tests)
                                        + explorer_refresh_spec.lua (5 tests)
   Same coverage; now runs as two parallel workers instead of one 34s
   serial file. Same before_each/setup_command scaffolding duplicated
   (deliberate — the two describes were self-contained already).

C. Consolidate 4 near-identical *_modules_spec.lua into one file:
   explorer_modules_spec.lua + lifecycle_modules_spec.lua +
   conflict_modules_spec.lua + history_modules_spec.lua
                                       -> tests/module_loading_spec.lua
   All 43 tests preserved; one place to keep the require-graph +
   public-API smoke in sync with the source.

D. Drop the _e2e_spec suffix on the recently-converted specs and
   normalize the issue-# convention:
   explorer_layout_e2e_spec.lua        -> explorer_window_layout_spec.lua
   explorer_toggle_e2e_spec.lua        -> explorer_visibility_spec.lua
   explorer_tree_e2e_spec.lua          -> explorer_tree_render_spec.lua
   history_layout_e2e_spec.lua         -> history_layout_spec.lua
   tab_cycle_untracked_e2e_spec.lua    -> issue_309_spec.lua
      (matches sibling issue_390_spec.lua / issue_496_spec.lua /
       issue_498_spec.lua naming in the same directory)
   Internal describe("… (E2E)") titles updated to match.

Verified: bash tests/run_tests.sh -> ALL TESTS PASSED (83 spec files,
28.4s), same 866-test surface as before this commit.
