+++ tests/keymap/registry_spec.lua
