+++ tests/module_loading_spec.lua
@@ -1,0 +1,366 @@
+-- Module loading smoke tests for every codediff.ui.* submodule.
+--
+-- Each folder has a set of internal Lua files plus an `init.lua` façade that
+-- re-exports them. The specs below verify:
+--   1) `require(module_path)` succeeds without erroring (catches syntax
+--      mistakes, missing files, circular-require breakage).
+--   2) The public API surface is preserved — each exported function name is
+--      still a function on the returned table (catches accidental removal
+--      of exports during refactors).
+--
+-- Consolidated from four separate `*_modules_spec.lua` files (explorer,
+-- lifecycle, conflict, history) because they were all the same shape:
+-- pcall(require) + assert.is_function on each export. One file is easier
+-- to locate, easier to keep in sync with the source, and slightly cheaper
+-- to run.
+
+-- ── explorer ────────────────────────────────────────────────────────────
+describe("Explorer submodules", function()
+  describe("module loading", function()
+    it("loads actions module", function()
+      local ok, mod = pcall(require, "codediff.ui.explorer.actions")
+      assert.is_true(ok, "Failed to require codediff.ui.explorer.actions: " .. tostring(mod))
+      assert.is_not_nil(mod)
+    end)
+
+    it("loads keymaps module", function()
+      local ok, mod = pcall(require, "codediff.ui.explorer.keymaps")
+      assert.is_true(ok, "Failed to require codediff.ui.explorer.keymaps")
+      assert.is_not_nil(mod)
+    end)
+
+    it("loads render module", function()
+      local ok, mod = pcall(require, "codediff.ui.explorer.render")
+      assert.is_true(ok, "Failed to require codediff.ui.explorer.render")
+      assert.is_not_nil(mod)
+    end)
+
+    it("loads refresh module", function()
+      local ok, mod = pcall(require, "codediff.ui.explorer.refresh")
+      assert.is_true(ok, "Failed to require codediff.ui.explorer.refresh")
+      assert.is_not_nil(mod)
+    end)
+
+    it("loads tree module", function()
+      local ok, mod = pcall(require, "codediff.ui.explorer.tree")
+      assert.is_true(ok, "Failed to require codediff.ui.explorer.tree")
+      assert.is_not_nil(mod)
+    end)
+
+    it("loads formatter and line rendering modules", function()
+      local formatters_ok, formatters = pcall(require, "codediff.ui.explorer.formatters")
+      local highlights_ok, highlights = pcall(require, "codediff.ui.explorer.line_highlights")
+      local layout_ok, layout = pcall(require, "codediff.ui.explorer.line_layout")
+      local stats_ok, stats = pcall(require, "codediff.ui.explorer.line_stats")
+      assert.is_true(formatters_ok, "Failed to require codediff.ui.explorer.formatters")
+      assert.is_true(highlights_ok, "Failed to require codediff.ui.explorer.line_highlights")
+      assert.is_true(layout_ok, "Failed to require codediff.ui.explorer.line_layout")
+      assert.is_true(stats_ok, "Failed to require codediff.ui.explorer.line_stats")
+      assert.is_not_nil(formatters)
+      assert.is_not_nil(highlights)
+      assert.is_not_nil(layout)
+      assert.is_not_nil(stats)
+    end)
+
+    it("loads init facade", function()
+      local ok, mod = pcall(require, "codediff.ui.explorer")
+      assert.is_true(ok, "Failed to require codediff.ui.explorer")
+      assert.is_not_nil(mod)
+    end)
+  end)
+
+  describe("public API", function()
+    it("actions exports expected functions", function()
+      local mod = require("codediff.ui.explorer.actions")
+      assert.is_function(mod.navigate_next)
+      assert.is_function(mod.navigate_prev)
+      assert.is_function(mod.toggle_visibility)
+      assert.is_function(mod.toggle_view_mode)
+      assert.is_function(mod.toggle_stage_file)
+      assert.is_function(mod.toggle_stage_entry)
+      assert.is_function(mod.stage_all)
+      assert.is_function(mod.unstage_all)
+      assert.is_function(mod.restore_entry)
+    end)
+
+    it("keymaps exports setup function", function()
+      local mod = require("codediff.ui.explorer.keymaps")
+      assert.is_function(mod.setup)
+    end)
+
+    it("render exports create function", function()
+      local mod = require("codediff.ui.explorer.render")
+      assert.is_function(mod.create)
+    end)
+
+    it("refresh exports expected functions", function()
+      local mod = require("codediff.ui.explorer.refresh")
+      assert.is_function(mod.setup_auto_refresh)
+      assert.is_function(mod.refresh)
+      assert.is_function(mod.get_all_files)
+    end)
+
+    it("tree exports create_tree_data", function()
+      local mod = require("codediff.ui.explorer.tree")
+      assert.is_function(mod.create_tree_data)
+    end)
+
+    it("formatters export callbacks for every explorer row type", function()
+      local mod = require("codediff.ui.explorer.formatters")
+      assert.is_function(mod.file)
+      assert.is_function(mod.folder)
+      assert.is_function(mod.group)
+    end)
+
+    it("line stats exports aggregation", function()
+      assert.is_function(require("codediff.ui.explorer.line_stats").sum)
+    end)
+  end)
+end)
+
+-- ── lifecycle ───────────────────────────────────────────────────────────
+describe("Lifecycle submodules", function()
+  describe("module loading", function()
+    it("loads session module", function()
+      local ok, mod = pcall(require, "codediff.ui.lifecycle.session")
+      assert.is_true(ok, "Failed to require codediff.ui.lifecycle.session")
+      assert.is_not_nil(mod)
+    end)
+
+    it("loads accessors module", function()
+      local ok, mod = pcall(require, "codediff.ui.lifecycle.accessors")
+      assert.is_true(ok, "Failed to require codediff.ui.lifecycle.accessors")
+      assert.is_not_nil(mod)
+    end)
+
+    it("loads state module", function()
+      local ok, mod = pcall(require, "codediff.ui.lifecycle.state")
+      assert.is_true(ok, "Failed to require codediff.ui.lifecycle.state")
+      assert.is_not_nil(mod)
+    end)
+
+    it("loads cleanup module", function()
+      local ok, mod = pcall(require, "codediff.ui.lifecycle.cleanup")
+      assert.is_true(ok, "Failed to require codediff.ui.lifecycle.cleanup")
+      assert.is_not_nil(mod)
+    end)
+
+    it("loads init facade without circular dependency", function()
+      local ok, mod = pcall(require, "codediff.ui.lifecycle")
+      assert.is_true(ok, "Failed to require codediff.ui.lifecycle - possible circular dependency")
+      assert.is_not_nil(mod)
+    end)
+  end)
+
+  describe("public API", function()
+    it("session exports expected functions", function()
+      local mod = require("codediff.ui.lifecycle.session")
+      assert.is_function(mod.get_active_diffs)
+      assert.is_function(mod.create_session)
+      assert.is_function(mod.compute_virtual_uri)
+    end)
+
+    it("accessors exports getter functions", function()
+      local mod = require("codediff.ui.lifecycle.accessors")
+      assert.is_function(mod.get_session)
+      assert.is_function(mod.get_mode)
+      assert.is_function(mod.get_git_context)
+      assert.is_function(mod.get_buffers)
+      assert.is_function(mod.get_windows)
+      assert.is_function(mod.get_paths)
+      assert.is_function(mod.find_tabpage_by_buffer)
+      assert.is_function(mod.is_original_virtual)
+      assert.is_function(mod.is_modified_virtual)
+      assert.is_function(mod.is_suspended)
+      assert.is_function(mod.get_explorer)
+      assert.is_function(mod.get_result)
+      assert.is_function(mod.get_conflict_blocks)
+      assert.is_function(mod.get_conflict_files)
+    end)
+
+    it("accessors exports setter functions", function()
+      local mod = require("codediff.ui.lifecycle.accessors")
+      assert.is_function(mod.update_suspended)
+      assert.is_function(mod.update_diff_result)
+      assert.is_function(mod.update_changedtick)
+      assert.is_function(mod.update_mtime)
+      assert.is_function(mod.update_paths)
+      assert.is_function(mod.update_buffers)
+      assert.is_function(mod.update_git_root)
+      assert.is_function(mod.update_revisions)
+      assert.is_function(mod.set_explorer)
+      assert.is_function(mod.set_result)
+      assert.is_function(mod.set_conflict_blocks)
+      assert.is_function(mod.track_conflict_file)
+    end)
+
+    it("state exports expected functions", function()
+      local mod = require("codediff.ui.lifecycle.state")
+      assert.is_function(mod.save_buffer_state)
+      assert.is_function(mod.restore_buffer_state)
+      assert.is_function(mod.clear_buffer_highlights)
+      assert.is_function(mod.get_file_mtime)
+      assert.is_function(mod.suspend_diff)
+      assert.is_function(mod.resume_diff)
+    end)
+
+    it("cleanup exports expected functions", function()
+      local mod = require("codediff.ui.lifecycle.cleanup")
+      assert.is_function(mod.setup_autocmds)
+      assert.is_function(mod.cleanup)
+      assert.is_function(mod.cleanup_all)
+      assert.is_function(mod.setup)
+    end)
+
+    it("init facade delegates all functions", function()
+      local mod = require("codediff.ui.lifecycle")
+      assert.is_function(mod.create_session)
+      assert.is_function(mod.cleanup)
+      assert.is_function(mod.cleanup_all)
+      assert.is_function(mod.setup)
+      assert.is_function(mod.get_session)
+      assert.is_function(mod.get_mode)
+      assert.is_function(mod.get_buffers)
+      assert.is_function(mod.get_windows)
+      assert.is_function(mod.set_explorer)
+    end)
+  end)
+
+  describe("circular dependency safety", function()
+    it("accessors can call get_active_diffs at runtime", function()
+      local accessors = require("codediff.ui.lifecycle.accessors")
+      -- Should not error - returns nil for non-existent tabpage
+      local result = accessors.get_session(99999)
+      assert.is_nil(result)
+    end)
+
+    it("state suspend_diff handles missing session gracefully", function()
+      local state = require("codediff.ui.lifecycle.state")
+      -- Should not error on non-existent tabpage
+      local ok = pcall(state.suspend_diff, 99999)
+      assert.is_true(ok, "suspend_diff should not error on missing session")
+    end)
+  end)
+end)
+
+-- ── conflict ────────────────────────────────────────────────────────────
+describe("Conflict submodules", function()
+  describe("module loading", function()
+    it("loads actions module", function()
+      local ok, mod = pcall(require, "codediff.ui.conflict.actions")
+      assert.is_true(ok, "Failed to require codediff.ui.conflict.actions")
+      assert.is_not_nil(mod)
+    end)
+
+    it("loads keymaps module", function()
+      local ok, mod = pcall(require, "codediff.ui.conflict.keymaps")
+      assert.is_true(ok, "Failed to require codediff.ui.conflict.keymaps")
+      assert.is_not_nil(mod)
+    end)
+
+    it("loads navigation module", function()
+      local ok, mod = pcall(require, "codediff.ui.conflict.navigation")
+      assert.is_true(ok, "Failed to require codediff.ui.conflict.navigation")
+      assert.is_not_nil(mod)
+    end)
+
+    it("loads signs module", function()
+      local ok, mod = pcall(require, "codediff.ui.conflict.signs")
+      assert.is_true(ok, "Failed to require codediff.ui.conflict.signs")
+      assert.is_not_nil(mod)
+    end)
+
+    it("loads diffget module", function()
+      local ok, mod = pcall(require, "codediff.ui.conflict.diffget")
+      assert.is_true(ok, "Failed to require codediff.ui.conflict.diffget")
+      assert.is_not_nil(mod)
+    end)
+
+    it("loads init facade", function()
+      local ok, mod = pcall(require, "codediff.ui.conflict")
+      assert.is_true(ok, "Failed to require codediff.ui.conflict")
+      assert.is_not_nil(mod)
+    end)
+  end)
+
+  describe("public API", function()
+    it("actions exports expected functions", function()
+      local mod = require("codediff.ui.conflict.actions")
+      assert.is_function(mod.accept_incoming)
+      assert.is_function(mod.accept_current)
+      assert.is_function(mod.accept_both)
+      assert.is_function(mod.discard)
+      assert.is_function(mod.accept_all_incoming)
+      assert.is_function(mod.accept_all_current)
+      assert.is_function(mod.accept_all_both)
+      assert.is_function(mod.discard_all)
+    end)
+
+    it("keymaps exports setup function", function()
+      local mod = require("codediff.ui.conflict.keymaps")
+      assert.is_function(mod.setup_keymaps)
+    end)
+
+    it("navigation exports expected functions", function()
+      local mod = require("codediff.ui.conflict.navigation")
+      assert.is_function(mod.navigate_next_conflict)
+      assert.is_function(mod.navigate_prev_conflict)
+    end)
+
+    it("signs exports expected functions", function()
+      local mod = require("codediff.ui.conflict.signs")
+      assert.is_function(mod.refresh_all_conflict_signs)
+      assert.is_function(mod.setup_sign_refresh_autocmd)
+    end)
+
+    it("diffget exports expected functions", function()
+      local mod = require("codediff.ui.conflict.diffget")
+      assert.is_function(mod.diffget_incoming)
+      assert.is_function(mod.diffget_current)
+    end)
+  end)
+end)
+
+-- ── history ─────────────────────────────────────────────────────────────
+describe("History submodules", function()
+  describe("module loading", function()
+    it("loads refresh module", function()
+      local ok, mod = pcall(require, "codediff.ui.history.refresh")
+      assert.is_true(ok, "Failed to require codediff.ui.history.refresh")
+      assert.is_not_nil(mod)
+    end)
+
+    it("loads render module", function()
+      local ok, mod = pcall(require, "codediff.ui.history.render")
+      assert.is_true(ok, "Failed to require codediff.ui.history.render")
+      assert.is_not_nil(mod)
+    end)
+
+    it("loads init facade", function()
+      local ok, mod = pcall(require, "codediff.ui.history")
+      assert.is_true(ok, "Failed to require codediff.ui.history")
+      assert.is_not_nil(mod)
+    end)
+  end)
+
+  describe("public API", function()
+    it("refresh exports expected functions", function()
+      local mod = require("codediff.ui.history.refresh")
+      assert.is_function(mod.setup_auto_refresh)
+      assert.is_function(mod.refresh)
+    end)
+
+    it("render exports expected functions", function()
+      local mod = require("codediff.ui.history.render")
+      assert.is_function(mod.build_tree_nodes)
+      assert.is_function(mod.create)
+      assert.is_function(mod.get_all_files)
+      assert.is_function(mod.navigate_next)
+      assert.is_function(mod.navigate_prev)
+      assert.is_function(mod.get_all_commits)
+      assert.is_function(mod.navigate_next_commit)
+      assert.is_function(mod.navigate_prev_commit)
+      assert.is_function(mod.toggle_visibility)
+    end)
+  end)
+end)
