--- tests/ui/conflict/conflict_modules_spec.lua
@@ -1,79 +1,0 @@
--- Module loading tests for conflict submodules
--- Validates that require() wiring works correctly after _set_*_module removal
-
-describe("Conflict submodules", function()
-  describe("module loading", function()
-    it("loads actions module", function()
-      local ok, mod = pcall(require, "codediff.ui.conflict.actions")
-      assert.is_true(ok, "Failed to require codediff.ui.conflict.actions")
-      assert.is_not_nil(mod)
-    end)
-
-    it("loads keymaps module", function()
-      local ok, mod = pcall(require, "codediff.ui.conflict.keymaps")
-      assert.is_true(ok, "Failed to require codediff.ui.conflict.keymaps")
-      assert.is_not_nil(mod)
-    end)
-
-    it("loads navigation module", function()
-      local ok, mod = pcall(require, "codediff.ui.conflict.navigation")
-      assert.is_true(ok, "Failed to require codediff.ui.conflict.navigation")
-      assert.is_not_nil(mod)
-    end)
-
-    it("loads signs module", function()
-      local ok, mod = pcall(require, "codediff.ui.conflict.signs")
-      assert.is_true(ok, "Failed to require codediff.ui.conflict.signs")
-      assert.is_not_nil(mod)
-    end)
-
-    it("loads diffget module", function()
-      local ok, mod = pcall(require, "codediff.ui.conflict.diffget")
-      assert.is_true(ok, "Failed to require codediff.ui.conflict.diffget")
-      assert.is_not_nil(mod)
-    end)
-
-    it("loads init facade", function()
-      local ok, mod = pcall(require, "codediff.ui.conflict")
-      assert.is_true(ok, "Failed to require codediff.ui.conflict")
-      assert.is_not_nil(mod)
-    end)
-  end)
-
-  describe("public API", function()
-    it("actions exports expected functions", function()
-      local mod = require("codediff.ui.conflict.actions")
-      assert.is_function(mod.accept_incoming)
-      assert.is_function(mod.accept_current)
-      assert.is_function(mod.accept_both)
-      assert.is_function(mod.discard)
-      assert.is_function(mod.accept_all_incoming)
-      assert.is_function(mod.accept_all_current)
-      assert.is_function(mod.accept_all_both)
-      assert.is_function(mod.discard_all)
-    end)
-
-    it("keymaps exports setup function", function()
-      local mod = require("codediff.ui.conflict.keymaps")
-      assert.is_function(mod.setup_keymaps)
-    end)
-
-    it("navigation exports expected functions", function()
-      local mod = require("codediff.ui.conflict.navigation")
-      assert.is_function(mod.navigate_next_conflict)
-      assert.is_function(mod.navigate_prev_conflict)
-    end)
-
-    it("signs exports expected functions", function()
-      local mod = require("codediff.ui.conflict.signs")
-      assert.is_function(mod.refresh_all_conflict_signs)
-      assert.is_function(mod.setup_sign_refresh_autocmd)
-    end)
-
-    it("diffget exports expected functions", function()
-      local mod = require("codediff.ui.conflict.diffget")
-      assert.is_function(mod.diffget_incoming)
-      assert.is_function(mod.diffget_current)
-    end)
-  end)
-end)
