--- tests/ui/explorer/explorer_modules_spec.lua
@@ -1,105 +1,0 @@
--- Module loading tests for explorer submodules
--- Validates that require() wiring works correctly after _set_*_module removal
-
-describe("Explorer submodules", function()
-  describe("module loading", function()
-    it("loads actions module", function()
-      local ok, mod = pcall(require, "codediff.ui.explorer.actions")
-      assert.is_true(ok, "Failed to require codediff.ui.explorer.actions: " .. tostring(mod))
-      assert.is_not_nil(mod)
-    end)
-
-    it("loads keymaps module", function()
-      local ok, mod = pcall(require, "codediff.ui.explorer.keymaps")
-      assert.is_true(ok, "Failed to require codediff.ui.explorer.keymaps")
-      assert.is_not_nil(mod)
-    end)
-
-    it("loads render module", function()
-      local ok, mod = pcall(require, "codediff.ui.explorer.render")
-      assert.is_true(ok, "Failed to require codediff.ui.explorer.render")
-      assert.is_not_nil(mod)
-    end)
-
-    it("loads refresh module", function()
-      local ok, mod = pcall(require, "codediff.ui.explorer.refresh")
-      assert.is_true(ok, "Failed to require codediff.ui.explorer.refresh")
-      assert.is_not_nil(mod)
-    end)
-
-    it("loads tree module", function()
-      local ok, mod = pcall(require, "codediff.ui.explorer.tree")
-      assert.is_true(ok, "Failed to require codediff.ui.explorer.tree")
-      assert.is_not_nil(mod)
-    end)
-
-    it("loads formatter and line rendering modules", function()
-      local formatters_ok, formatters = pcall(require, "codediff.ui.explorer.formatters")
-      local highlights_ok, highlights = pcall(require, "codediff.ui.explorer.line_highlights")
-      local layout_ok, layout = pcall(require, "codediff.ui.explorer.line_layout")
-      local stats_ok, stats = pcall(require, "codediff.ui.explorer.line_stats")
-      assert.is_true(formatters_ok, "Failed to require codediff.ui.explorer.formatters")
-      assert.is_true(highlights_ok, "Failed to require codediff.ui.explorer.line_highlights")
-      assert.is_true(layout_ok, "Failed to require codediff.ui.explorer.line_layout")
-      assert.is_true(stats_ok, "Failed to require codediff.ui.explorer.line_stats")
-      assert.is_not_nil(formatters)
-      assert.is_not_nil(highlights)
-      assert.is_not_nil(layout)
-      assert.is_not_nil(stats)
-    end)
-
-    it("loads init facade", function()
-      local ok, mod = pcall(require, "codediff.ui.explorer")
-      assert.is_true(ok, "Failed to require codediff.ui.explorer")
-      assert.is_not_nil(mod)
-    end)
-  end)
-
-  describe("public API", function()
-    it("actions exports expected functions", function()
-      local mod = require("codediff.ui.explorer.actions")
-      assert.is_function(mod.navigate_next)
-      assert.is_function(mod.navigate_prev)
-      assert.is_function(mod.toggle_visibility)
-      assert.is_function(mod.toggle_view_mode)
-      assert.is_function(mod.toggle_stage_file)
-      assert.is_function(mod.toggle_stage_entry)
-      assert.is_function(mod.stage_all)
-      assert.is_function(mod.unstage_all)
-      assert.is_function(mod.restore_entry)
-    end)
-
-    it("keymaps exports setup function", function()
-      local mod = require("codediff.ui.explorer.keymaps")
-      assert.is_function(mod.setup)
-    end)
-
-    it("render exports create function", function()
-      local mod = require("codediff.ui.explorer.render")
-      assert.is_function(mod.create)
-    end)
-
-    it("refresh exports expected functions", function()
-      local mod = require("codediff.ui.explorer.refresh")
-      assert.is_function(mod.setup_auto_refresh)
-      assert.is_function(mod.refresh)
-      assert.is_function(mod.get_all_files)
-    end)
-
-    it("tree exports create_tree_data", function()
-      local mod = require("codediff.ui.explorer.tree")
-      assert.is_function(mod.create_tree_data)
-    end)
-
-    it("formatters export callbacks for every explorer row type", function()
-      local mod = require("codediff.ui.explorer.formatters")
-      assert.is_function(mod.file)
-      assert.is_function(mod.folder)
-      assert.is_function(mod.group)
-    end)
-
-    it("line stats exports aggregation", function()
-      assert.is_function(require("codediff.ui.explorer.line_stats").sum)
-    end)
-  end)
-end)
