--- tests/ui/lifecycle/lifecycle_modules_spec.lua
@@ -1,127 +1,0 @@
--- Module loading tests for lifecycle submodules
--- Validates that require() wiring works correctly after _set_*_module removal
--- Special attention to session ↔ accessors circular dependency
-
-describe("Lifecycle submodules", function()
-  describe("module loading", function()
-    it("loads session module", function()
-      local ok, mod = pcall(require, "codediff.ui.lifecycle.session")
-      assert.is_true(ok, "Failed to require codediff.ui.lifecycle.session")
-      assert.is_not_nil(mod)
-    end)
-
-    it("loads accessors module", function()
-      local ok, mod = pcall(require, "codediff.ui.lifecycle.accessors")
-      assert.is_true(ok, "Failed to require codediff.ui.lifecycle.accessors")
-      assert.is_not_nil(mod)
-    end)
-
-    it("loads state module", function()
-      local ok, mod = pcall(require, "codediff.ui.lifecycle.state")
-      assert.is_true(ok, "Failed to require codediff.ui.lifecycle.state")
-      assert.is_not_nil(mod)
-    end)
-
-    it("loads cleanup module", function()
-      local ok, mod = pcall(require, "codediff.ui.lifecycle.cleanup")
-      assert.is_true(ok, "Failed to require codediff.ui.lifecycle.cleanup")
-      assert.is_not_nil(mod)
-    end)
-
-    it("loads init facade without circular dependency", function()
-      local ok, mod = pcall(require, "codediff.ui.lifecycle")
-      assert.is_true(ok, "Failed to require codediff.ui.lifecycle - possible circular dependency")
-      assert.is_not_nil(mod)
-    end)
-  end)
-
-  describe("public API", function()
-    it("session exports expected functions", function()
-      local mod = require("codediff.ui.lifecycle.session")
-      assert.is_function(mod.get_active_diffs)
-      assert.is_function(mod.create_session)
-      assert.is_function(mod.compute_virtual_uri)
-    end)
-
-    it("accessors exports getter functions", function()
-      local mod = require("codediff.ui.lifecycle.accessors")
-      assert.is_function(mod.get_session)
-      assert.is_function(mod.get_mode)
-      assert.is_function(mod.get_git_context)
-      assert.is_function(mod.get_buffers)
-      assert.is_function(mod.get_windows)
-      assert.is_function(mod.get_paths)
-      assert.is_function(mod.find_tabpage_by_buffer)
-      assert.is_function(mod.is_original_virtual)
-      assert.is_function(mod.is_modified_virtual)
-      assert.is_function(mod.is_suspended)
-      assert.is_function(mod.get_explorer)
-      assert.is_function(mod.get_result)
-      assert.is_function(mod.get_conflict_blocks)
-      assert.is_function(mod.get_conflict_files)
-    end)
-
-    it("accessors exports setter functions", function()
-      local mod = require("codediff.ui.lifecycle.accessors")
-      assert.is_function(mod.update_suspended)
-      assert.is_function(mod.update_diff_result)
-      assert.is_function(mod.update_changedtick)
-      assert.is_function(mod.update_mtime)
-      assert.is_function(mod.update_paths)
-      assert.is_function(mod.update_buffers)
-      assert.is_function(mod.update_git_root)
-      assert.is_function(mod.update_revisions)
-      assert.is_function(mod.set_explorer)
-      assert.is_function(mod.set_result)
-      assert.is_function(mod.set_conflict_blocks)
-      assert.is_function(mod.track_conflict_file)
-    end)
-
-    it("state exports expected functions", function()
-      local mod = require("codediff.ui.lifecycle.state")
-      assert.is_function(mod.save_buffer_state)
-      assert.is_function(mod.restore_buffer_state)
-      assert.is_function(mod.clear_buffer_highlights)
-      assert.is_function(mod.get_file_mtime)
-      assert.is_function(mod.suspend_diff)
-      assert.is_function(mod.resume_diff)
-    end)
-
-    it("cleanup exports expected functions", function()
-      local mod = require("codediff.ui.lifecycle.cleanup")
-      assert.is_function(mod.setup_autocmds)
-      assert.is_function(mod.cleanup)
-      assert.is_function(mod.cleanup_all)
-      assert.is_function(mod.setup)
-    end)
-
-    it("init facade delegates all functions", function()
-      local mod = require("codediff.ui.lifecycle")
-      assert.is_function(mod.create_session)
-      assert.is_function(mod.cleanup)
-      assert.is_function(mod.cleanup_all)
-      assert.is_function(mod.setup)
-      assert.is_function(mod.get_session)
-      assert.is_function(mod.get_mode)
-      assert.is_function(mod.get_buffers)
-      assert.is_function(mod.get_windows)
-      assert.is_function(mod.set_explorer)
-    end)
-  end)
-
-  describe("circular dependency safety", function()
-    it("accessors can call get_active_diffs at runtime", function()
-      local accessors = require("codediff.ui.lifecycle.accessors")
-      -- Should not error - returns nil for non-existent tabpage
-      local result = accessors.get_session(99999)
-      assert.is_nil(result)
-    end)
-
-    it("state suspend_diff handles missing session gracefully", function()
-      local state = require("codediff.ui.lifecycle.state")
-      -- Should not error on non-existent tabpage
-      local ok = pcall(state.suspend_diff, 99999)
-      assert.is_true(ok, "suspend_diff should not error on missing session")
-    end)
-  end)
-end)
